Add issues filtering

+++ lib/AWSomeLink/R.pm
@@ -51,8 +51,25 @@ sub issues{
 
     my $prefix = Utils::trim $c->stash->{prefix} ;
     my $db = Db->new($c,$prefix) ;
-    Utils::deploy_db_objects($c,$db,'issue','issues') ;
+    Utils::deploy_db_objects($c,$db,'issue','issues',$c->session('issues filter')) ;
     Utils::deploy_db_objects($c,$db,'recipient','participants') ;
+};
+
+sub issues_nofilter{
+    my $c = shift ;
+    return if ! _check_access($c);
+
+    my $prefix = Utils::trim $c->stash->{prefix} ;
+    $c->session( 'issues filter' => undef );
+    $c->redirect_to( "/$prefix/r/issues" ) ;
+};
+
+sub issues_filter{
+    my $c = shift ;
+    return if ! _check_access($c);
+
+    $c->session( 'issues filter' => $c->param('filter') );
+    $c->redirect_to($c->param('return_path')) ;
 };
 
 sub issues_edit{
@@ -63,7 +80,13 @@ sub issues_edit{
     my $db = Db->new($c,$prefix) ;
     my $payload = Utils::trim $c->stash->{payload};
     if( $payload ){
-        Utils::R::issue_edit($c,$db,$payload) if lc($c->req->method) eq 'post' ;
+        if( lc($c->req->method) eq 'post' ){
+            if( Utils::R::check_rw_access($db,$payload,$c->session('user id')) ){
+                Utils::R::issue_edit($c,$db,$payload) ;
+            } else {
+                $c->stash( error_rw => 1 );
+            }
+        }
         Utils::deploy_db_object($c,$db,$payload) ;
         Utils::deploy_db_objects($c,$db,'property','properties') ;
     } else {
@@ -94,12 +117,15 @@ sub issues_delete_files{
     my $prefix = Utils::trim $c->stash->{prefix} ;
     my $payload = Utils::trim $c->stash->{payload};
 
-    my $files = $c->every_param('files');
-    my $path = Utils::R::get_files_path($c,$payload);
-    for my $file_name (@{$files}){
-        my $file_path = "$path/$file_name";
-        unlink $file_path if -e $file_path ;
-    }
+    my $db = Db->new($c,$prefix) ;
+    if( Utils::R::check_rw_access($db,$payload,$c->session('user id')) ){
+        my $files = $c->every_param('files');
+        my $path = Utils::R::get_files_path($c,$payload);
+        for my $file_name (@{$files}){
+            my $file_path = "$path/$file_name";
+            unlink $file_path if -e $file_path ;
+        }
+    }   
     $c->redirect_to("/$prefix/r/issues_edit/$payload");
 };
 
@@ -111,11 +137,13 @@ sub issues_delete {
     my $db = Db->new($c,$prefix) ;
     my $payload = Utils::trim $c->stash->{payload};
 
-    # 1. delete db record
-    $db->del($payload) ;
-    # 2. delete files # !!!VERY DANGEROUS!!!
-    my $path = Utils::R::get_files_path($c,$payload);
-    system("rm -fR $path");
+    if( Utils::R::check_rw_access($db,$payload,$c->session('user id')) ){
+        # 1. delete db record
+        $db->del($payload) ;
+        # 2. delete files # !!!VERY DANGEROUS!!!
+        my $path = Utils::R::get_files_path($c,$payload);
+        system("rm -fR $path");
+    }
 
     $c->redirect_to("/$prefix/r/issues");
 };
